src/test/java/com/example/demo/converters/IngredientCommandToIngredientTest.java
package com.example.demo.converters;

import static org.junit.Assert.assertEquals;
import static org.junit.Assert.assertNotNull;
import static org.junit.Assert.assertNull;

import java.math.BigDecimal;

import org.junit.Before;
import org.junit.Test;

import com.example.demo.converter.IngredientCommandToIngredient;
import com.example.demo.converter.StringToUomConverters;
import com.example.demo.entities.Ingredient;
import com.example.demo.entities.Recipe;
import com.example.demo.entities.UnitOfMeasure;
import com.example.demo.form.IngredientCommand;

public class IngredientCommandToIngredientTest {

    public static final Recipe RECIPE = new Recipe();
    public static final BigDecimal AMOUNT = new BigDecimal("1");
    public static final String NAME = "ingredient";
    public static final Long ID_VALUE = 1L;
    public static final UnitOfMeasure UOM = UnitOfMeasure.CUP;
    
    StringToUomConverters stringToUomConverters;

    IngredientCommandToIngredient converter;
    

    @Before
    public void setUp() throws Exception {
        converter = new IngredientCommandToIngredient(new StringToUomConverters());
    }

    @Test
    public void testNullObject() throws Exception {
        assertNull(converter.convert(null));
    }

    @Test
    public void testEmptyObject() throws Exception {
        assertNotNull(converter.convert(new IngredientCommand()));
    }

    @Test
    public void convert() throws Exception {
        //given
        IngredientCommand command = new IngredientCommand();
        command.setId(ID_VALUE);
        command.setAmount(AMOUNT);
        command.setName(NAME);
        command.setUom(UOM.getValue());

        //when
        Ingredient ingredient = converter.convert(command);

        //then
        assertNotNull(ingredient);
        assertNotNull(ingredient.getUom());
        assertEquals(ID_VALUE, ingredient.getId());
        assertEquals(AMOUNT, ingredient.getAmount());
        assertEquals(NAME, ingredient.getName());
        assertEquals(UOM, ingredient.getUom());
    }

    @Test
    public void convertWithNullUOM() throws Exception {
        //given
        IngredientCommand command = new IngredientCommand();
        command.setId(ID_VALUE);
        command.setAmount(AMOUNT);
        command.setName(NAME);


        //when
        Ingredient ingredient = converter.convert(command);

        //then
        assertNotNull(ingredient);
        assertNull(ingredient.getUom());
        assertEquals(ID_VALUE, ingredient.getId());
        assertEquals(AMOUNT, ingredient.getAmount());
        assertEquals(NAME, ingredient.getName());
    }

}